chore: bump to stable release version `v1.0.11` of fish-lsp
* `fish-lsp@latest` tag is bumped to `fish-lsp@1.0.11`

* man file is located at `man/man1/fish-lsp.1` instead of
`man/fish-lsp.1` since the previous man folder structure depends on the
package manager's feature set to install man/fish-lsp.1 in the correct
location. `man/man1/fish-lsp.1` relies less heavily on the package
manager to just install the man directory into the global package
manager's equivalent location

* updated `man/man1/fish-lsp.1` to point to correct locations in `src/*`
&& `dist/fish-lsp` embedded output. Our changes to the source code allow
the virtual-fs shipped in the embedded bundle to resolve to the correct
new location. We also updated `scripts/*` that mess with the manpage and
lastly we editted the `generate:man` script in `package.json`, to
redirect its output to the `man/man1/fish-lsp.1` file inside the local
project

* updated `out/build-time.json` lookup utilities, which is built during
installation of package via package manager, or embedded directly into
the server. This allows the binary to keep track of when the binary was
installed from a command like: `npm i -g fish-lsp@latest`. If no
out/package.json file is found, it fallsback to an embedded version of
the file when the binary was built to be published.