refactor(`LspDocument`): added static `LspDocument.is{Funced,CommandlineBuffer}Path()` utils
- kept `edit_command_buffer` support, but simplified how `isFuncedPath`
and `isCommandlineBuffer` can be used by defining their utilities as
static public methods on `LspDocument.

- we also we now allowed to treat `LspDocument.isFuncedPath(path) ||
LspDocument.isCommandlineBuffer(path)` the same in many locations since
we have extracted the autoloaded type of a document to the `LspDocument`
class, and both of our extra cases are used to return
`$__fish_config_dir` workspace.

- cleaned up unused functions